Implementación del sistema difuso de alerta temprana para inundaciones
- Agregada la lógica de defuzzificación en defuzzification.py con métodos de centroide y centro de máximos.
- Creación del motor de inferencia en engine.py, que incluye fuzzificación, evaluación de reglas, implicación y agregación.
- Definición de funciones de membresía en membership.py utilizando funciones trapezoidales.
- Implementación de funciones de visualización en plotting.py para graficar variables, conjuntos implicados y resultados de agregación.
- Definición de reglas difusas en rules.py para el sistema de alerta.
- Configuración de variables de entrada y salida en variables.py, incluyendo niveles de río y precipitación.
- Scripts de ejecución (run.bat y run.sh) para facilitar la instalación y ejecución del sistema.

+++ engine.py
@@ -1,7 +1,7 @@
 import numpy as np
 from dataclasses import dataclass
-from .membership import LinguisticVariable, MembershipFunction
-from .rules import FuzzyRule, Operator
+from membership import MembershipFunction
+from rules import FuzzyRule, Operator
 
 
 @dataclass
@@ -76,7 +76,7 @@ def run_inference(input_variables, output_variable, rules, crisp_inputs,
     activated = _evaluate_rules(rules, fuzzified, output_variable)
 
     u_min, u_max = output_variable.universe
-    n = int(round((u_max - u_min) / resolution)) + 1
+    n = round((u_max - u_min) / resolution) + 1
     y = np.linspace(u_min, u_max, n)
 
     implied_sets = []
